requirements.txt
cairocffi==1.7.1
cairosvg==2.7.1
cffi==1.17.1
contourpy==1.3.1
cssselect2==0.8.0
cycler==0.12.1
defusedxml==0.7.1
fonttools==4.56.0
kids-cache==0.0.7
kiwisolver==1.4.8
libsass==0.23.0
lxml==5.3.1
matplotlib==3.10.1
mock==5.2.0
numpy==2.2.4
packaging==24.2
perlin-noise==1.13
pillow==11.1.0
pycparser==2.22
pyparsing==3.2.3
pyqt6==6.4.2
pyqt6-qt6==6.8.2
pyqt6-sip==13.10.0
pyside6==6.8.3
pyside6-addons==6.8.3
pyside6-essentials==6.8.3
python-dateutil==2.9.0.post0
qt-pyqt-pyside-custom-widgets==1.0.2
qtpy==2.4.3
qtsass==0.4.0
setuptools==78.1.0
shiboken6==6.8.3
six==1.17.0
termcolor==3.0.1
tinycss2==1.4.0
watchdog==6.0.0
webencodings==0.5.1
